[MISC] Rename {core->utlity}/char_operations/predicate.hpp.
