+++ include/seqan3/argument_parser/detail/format_parse.hpp
@@ -19,7 +19,7 @@
 #include <vector>
 
 #include <seqan3/argument_parser/detail/format_base.hpp>
-#include <seqan3/core/char_operations/predicate.hpp>
+#include <seqan3/utility/char_operations/predicate.hpp>
 #include <seqan3/utility/detail/type_name_as_string.hpp>
 
 namespace seqan3::detail
