+++ include/seqan3/io/alignment_file/format_bam.hpp
@@ -21,7 +21,6 @@
 #include <seqan3/alphabet/detail/convert.hpp>
 #include <seqan3/alphabet/nucleotide/sam_dna16.hpp>
 #include <seqan3/core/bit_manipulation.hpp>
-#include <seqan3/core/char_operations/predicate.hpp>
 #include <seqan3/core/concept/core_language.hpp>
 #include <seqan3/core/concept/tuple.hpp>
 #include <seqan3/core/debug_stream/detail/to_string.hpp>
@@ -44,6 +43,7 @@
 #include <seqan3/range/views/slice.hpp>
 #include <seqan3/range/views/take_exactly.hpp>
 #include <seqan3/range/views/take_until.hpp>
+#include <seqan3/utility/char_operations/predicate.hpp>
 
 namespace seqan3
 {
