+++ include/seqan3/io/alignment_file/format_sam.hpp
@@ -19,7 +19,6 @@
 #include <string>
 #include <vector>
 
-#include <seqan3/core/char_operations/predicate.hpp>
 #include <seqan3/core/concept/core_language.hpp>
 #include <seqan3/core/concept/tuple.hpp>
 #include <seqan3/core/debug_stream/detail/to_string.hpp>
@@ -46,6 +45,7 @@
 #include <seqan3/range/views/take_until.hpp>
 #include <seqan3/range/views/to_char.hpp>
 #include <seqan3/range/views/to.hpp>
+#include <seqan3/utility/char_operations/predicate.hpp>
 
 namespace seqan3
 {
