+++ include/seqan3/io/alignment_file/sam_tag_dictionary.hpp
@@ -12,14 +12,14 @@
 
 #pragma once
 
+#include <seqan3/std/concepts>
 #include <map>
 #include <variant>
 
-#include <seqan3/core/char_operations/predicate.hpp>
 #include <seqan3/core/concept/core_language.hpp>
 #include <seqan3/core/type_traits/template_inspection.hpp>
 #include <seqan3/range/container/small_string.hpp>
-#include <seqan3/std/concepts>
+#include <seqan3/utility/char_operations/predicate.hpp>
 
 namespace seqan3::detail
 {
