+++ include/seqan3/io/sequence_file/format_embl.hpp
@@ -12,13 +12,14 @@
 
 #pragma once
 
+#include <seqan3/std/algorithm>
 #include <iterator>
+#include <seqan3/std/ranges>
 #include <string>
 #include <string_view>
 #include <vector>
 
 #include <seqan3/alphabet/nucleotide/dna5.hpp>
-#include <seqan3/core/char_operations/predicate.hpp>
 #include <seqan3/core/type_traits/range.hpp>
 #include <seqan3/io/detail/misc.hpp>
 #include <seqan3/io/sequence_file/input_format_concept.hpp>
@@ -33,8 +34,7 @@
 #include <seqan3/range/views/to_char.hpp>
 #include <seqan3/range/views/take_line.hpp>
 #include <seqan3/range/views/take_until.hpp>
-#include <seqan3/std/algorithm>
-#include <seqan3/std/ranges>
+#include <seqan3/utility/char_operations/predicate.hpp>
 
 namespace seqan3
 {
