+++ include/seqan3/range/views/take_line.hpp
@@ -12,8 +12,8 @@
 
 #pragma once
 
-#include <seqan3/core/char_operations/predicate.hpp>
 #include <seqan3/range/views/take_until.hpp>
+#include <seqan3/utility/char_operations/predicate.hpp>
 
 // ============================================================================
 //  views::take_line (adaptor instance definition)
