+++ include/seqan3/utility/char_operations/all.hpp
@@ -13,11 +13,9 @@
 
 #pragma once
 
-#include <seqan3/core/char_operations/pretty_print.hpp>
-#include <seqan3/core/char_operations/transform.hpp>
-#include <seqan3/utility/char_operations/all.hpp>
+#include <seqan3/utility/char_operations/predicate.hpp>
 
 /*!\defgroup char_operations builtin_character operations
  * \brief Provides various operations on character types.
- * \ingroup core
+ * \ingroup utility
  */
