+++ test/performance/utility/char_operations/char_predicate_benchmark.cpp
@@ -11,8 +11,8 @@
 
 #include <benchmark/benchmark.h>
 
-#include <seqan3/core/char_operations/predicate.hpp>
 #include <seqan3/test/seqan2.hpp>
+#include <seqan3/utility/char_operations/predicate.hpp>
 
 #if SEQAN3_HAS_SEQAN2
 #include <seqan/stream.h>
