+++ test/snippet/utility/char_operations/char_predicate.cpp
@@ -1,5 +1,5 @@
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
-#include <seqan3/core/char_operations/predicate.hpp>
+#include <seqan3/utility/char_operations/predicate.hpp>
 
 int main()
 {
