+++ test/snippet/utility/char_operations/char_predicate_is_char.cpp
@@ -1,4 +1,4 @@
-#include <seqan3/core/char_operations/predicate.hpp>
+#include <seqan3/utility/char_operations/predicate.hpp>
 
 int main()
 {
