+++ test/unit/alphabet/nucleotide/rna15_test.cpp
@@ -6,8 +6,8 @@
 // -----------------------------------------------------------------------------------------------------
 
 #include <seqan3/alphabet/nucleotide/rna15.hpp>
-#include <seqan3/core/char_operations/predicate.hpp>
 #include <seqan3/core/detail/debug_stream_range.hpp>
+#include <seqan3/utility/char_operations/predicate.hpp>
 
 #include "../alphabet_constexpr_test_template.hpp"
 #include "../alphabet_test_template.hpp"
