+++ test/unit/range/views/istreambuf_test.cpp
@@ -12,7 +12,6 @@
 #include <seqan3/std/ranges>
 
 #include <seqan3/alphabet/nucleotide/dna5.hpp>
-#include <seqan3/core/char_operations/predicate.hpp>
 #include <seqan3/core/detail/debug_stream_alphabet.hpp>
 #include <seqan3/range/views/char_to.hpp>
 #include <seqan3/range/views/complement.hpp>
@@ -21,6 +20,7 @@
 #include <seqan3/range/views/to.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
 #include <seqan3/test/tmp_filename.hpp>
+#include <seqan3/utility/char_operations/predicate.hpp>
 
 #include "../iterator_test_template.hpp"
